编辑物品

+++ Shio Valley/Assets/Scripts/Data/Item/ItemDataManager.cs
@@ -3,6 +3,7 @@ using UnityEngine;
 using System.Collections.Generic;
 using UnityEngine.AddressableAssets;
 using System.Threading.Tasks;
+using System.Runtime.InteropServices.WindowsRuntime;
 
 public static class ItemDataManager
 {
@@ -16,6 +17,24 @@ public static class ItemDataManager
         if (!File.Exists(filePath))
         {
             Debug.LogWarning("Item database not found, creating new one.");
+            string path;
+#if UNITY_EDITOR
+            path = Path.Combine(Application.dataPath, "IniData/items.json");
+            string _json = File.ReadAllText(path);
+            ItemDatabase _itemDatabase = JsonUtility.FromJson<ItemDatabase>(_json);
+            foreach (Item item in _itemDatabase.items)
+            {
+                if (item.canUsedinseason == null || item.canUsedinseason.Count != 4)
+                {
+                    item.canUsedinseason = new List<bool>();
+                    for (int i = 0; i < 4; i++)
+                    {
+                        item.canUsedinseason.Add(false);
+                    }
+                }
+            }
+            return _itemDatabase;
+#endif
             return new ItemDatabase { items = new List<Item>() };
         }
         string json = File.ReadAllText(filePath);
@@ -36,6 +55,7 @@ public static class ItemDataManager
     public static void Save(ItemDatabase database)
     {
         string json = JsonUtility.ToJson(database);
+        if(File.Exists(filePath))
         File.WriteAllText(filePath, json);
 #if UNITY_EDITOR
         string iniFilePath = Path.Combine(Application.dataPath, "IniData/items.json");
